Add Menu module and button event listeners
Similar to the home page content, the menu page needs its own js
file to control the content of the menu. This module is imported
into index.js and the loadMenu function is called when the Menu
button is clicked. This function is thus called in the Menu button's
event listener. Additionally, an event listener has been added for
the home button so that the user can return to the home menu.